Semver is back (#1366)
* semver is back

* should be good now
