+++ .github/workflows/_component_prechecks.yml
@@ -1,0 +1,46 @@
+name: nucliadb Client (py)
+
+on:
+  workflow_call:
+    inputs:
+      component:
+        required: true
+        type: string
+
+
+jobs:
+  # Job to run pre-checks
+  pre_checks:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.9", "3.10", "3.11"]
+
+    steps:
+      - name: Checkout the repository
+        uses: actions/checkout@v3
+
+      - name: Setup Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+          cache: "pip"
+
+      - name: Install package
+        run: make -C ${{ inputs.component }}/ install-dev
+
+      - name: Run pre-checks
+        run: make -C ${{ inputs.component }}/ lint
+      
+      - uses: dorny/paths-filter@v2
+        id: changes
+        with:
+          filters: |
+            VERSION:
+              - '${{ inputs.component }}/VERSION'
+      
+      - name: Check if VERSION file was not updated
+        if: steps.changes.outputs.VERSION != 'true'
+        run: |
+          echo "VERSION file was not updated"
+          exit 1
