+++ .github/workflows/nucliadb_models.yml
@@ -13,25 +13,7 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  # Job to run pre-checks
   pre-checks:
-    name: NucliaDBModelsTests
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ["3.9", "3.10", "3.11"]
-
-    steps:
-      - name: Checkout the repository
-        uses: actions/checkout@v2
-
-      - name: Setup Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-
-      - name: Install package
-        run: make -C nucliadb_models/ install-dev
-
-      - name: Run pre-checks
-        run: make -C nucliadb_models/ lint
+    uses: ./.github/workflows/_component_prechecks.yml
+    with:
+      component: "nucliadb_models"
