+++ .github/workflows/nucliadb_telemetry.yml
@@ -19,33 +19,14 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  # Job to run pre-checks
   pre-checks:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ["3.9", "3.10", "3.11"]
-
-    steps:
-      - name: Checkout the repository
-        uses: actions/checkout@v3
-
-      - name: Setup Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-          cache: "pip"
-
-      - name: Install package
-        run: |
-          make -C nucliadb_telemetry/ install-dev
-
-      - name: Run pre-checks
-        run: |
-          make -C nucliadb_telemetry/ lint
+    uses: ./.github/workflows/_component_prechecks.yml
+    with:
+      component: "nucliadb_telemetry"
 
   # Job to run tests
   tests:
+    name: NucliaDBTelemetryTests
     runs-on: ubuntu-latest
 
     strategy:
