--- .github/workflows/nucliadb_telemetry_release.yml
@@ -1,41 +1,0 @@
-name: Release NucliaDB Telemetry
-
-on:
-  push:
-    branches:
-      - main
-    paths:
-      - ".github/workflows/nucliadb_telemetry_release.yml"
-      - "nucliadb_telemetry/VERSION"
-
-jobs:
-  build_wheels:
-    name: Build wheels
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          submodules: recursive
-          fetch-depth: 0
-
-      - uses: actions/setup-python@v4
-        with:
-          python-version: 3.11
-
-      - name: Install requirements
-        run: |
-          pip install -r nucliadb_telemetry/requirements.txt
-          pip install wheel
-
-      - name: Install nucliadb_telemetry
-        run: pip install -e ./nucliadb_telemetry
-
-      - name: Build wheels
-        run: pip wheel --no-deps -w dist ./nucliadb_telemetry
-
-      - name: Publish distribution 📦 to PyPI
-        uses: pypa/gh-action-pypi-publish@release/v1
-        with:
-          user: __token__
-          password: ${{ secrets.PYPI_TOKEN }}
