+++ .github/workflows/nucliadb_utils.yml
@@ -15,26 +15,9 @@ concurrency:
 jobs:
   # Job to run pre-checks
   pre-checks:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ["3.9", "3.10", "3.11"]
-
-    steps:
-      - name: Checkout the repository
-        uses: actions/checkout@v3
-
-      - name: Setup Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-          cache: "pip"
-
-      - name: Install package
-        run: make -C nucliadb_utils/ install-dev
-
-      - name: Run pre-checks
-        run: make -C nucliadb_utils/ lint
+    uses: ./.github/workflows/_component_prechecks.yml
+    with:
+      component: "nucliadb_utils"
 
   # Job to run tests
   tests:
