+++ bump.py
@@ -5,10 +5,13 @@ parser.add_argument("--build", type=int, help="build number")
 parser.add_argument("--sem", type=str, help="Semantic version part")
 parser.add_argument("--minor", type=bool, help="Minor")
 parser.add_argument("--bug", type=bool, help="Bug")
+parser.add_argument(
+    "--version_file", type=str, help="Version file to update", default="VERSION"
+)
 
 
 def run(args):
-    with open("VERSION", "r") as f:
+    with open(args.version_file, "r") as f:
         version = f.read().strip()
 
     major, minor, bug = version.split(".")
@@ -30,7 +33,7 @@ def run(args):
         bug += 1
 
     version = f"{major}.{minor}.{bug}{version_post}"
-    with open("VERSION", "w") as f:
+    with open(args.version_file, "w") as f:
         f.write(version)
