+++ nucliadb/nucliadb/tests/unit/test_openapi.py
@@ -21,8 +21,8 @@ from unittest.mock import Mock
 
 from starlette.routing import Mount
 
+from nucliadb.openapi import extract_openapi, is_versioned_route
 from nucliadb.search.app import application
-from nucliadb.search.openapi import extract_openapi, is_versioned_route
 
 
 def get_route(path):
@@ -35,4 +35,4 @@ def test_is_versioned_route():
 
 
 def test_extract_openapi():
-    assert extract_openapi(application, "1", "commitid")
+    assert extract_openapi(application, "1", "commitid", "nucliadb_search")
