+++ nucliadb/requirements.txt
@@ -38,9 +38,9 @@ aiodns>=3.0.0
 
 types-orjson
 nucliadb-telemetry>=1.15.0
-nucliadb-utils[storages,fastapi,cache]
-nucliadb-protos
-nucliadb-models>=2.15.1.post353
+nucliadb-utils[storages,fastapi,cache]>=2.24.0
+nucliadb-protos>=2.24.0
+nucliadb-models>=2.24.0
 nucliadb-contributor-assets~=1.0.0
 
 asyncpg>=0.27.0
