+++ nucliadb_client/requirements-source.txt
@@ -1,4 +1,3 @@
 -e ./nucliadb_protos/python
 -e ./nucliadb_models
 -e ./nucliadb_telemetry
--e ./nucliadb_utils
