+++ nucliadb_client/requirements.txt
@@ -3,7 +3,7 @@ protobuf>=3.19.3
 types-protobuf>=3.19.5,<4.0
 grpcio>=1.44.0
 aiofiles
-nucliadb-protos
-nucliadb-models
+nucliadb-protos>=2.24.0
+nucliadb-models>=2.24.0
 nucliadb-telemetry>=1.15.0
 urllib3<1.27,>=1.21.1
