+++ nucliadb_dataset/VERSION
@@ -1,0 +1,1 @@
+2.24.0
