+++ nucliadb_dataset/setup.py
@@ -4,17 +4,14 @@ from pathlib import Path
 from setuptools import find_packages, setup  # type: ignore
 
 _dir = Path(__file__).resolve().parent
-VERSION = _dir.parent.joinpath("VERSION").open().read().strip()
+VERSION = _dir.joinpath("VERSION").open().read().strip()
 README = _dir.joinpath("README.md").open().read()
 
 
 def load_reqs(filename):
     with open(filename) as reqs_file:
         return [
-            # pin nucliadb-xxx to the same version as nucliadb
-            line.strip() + f"=={VERSION}"
-            if line.startswith("nucliadb-") and "=" not in line
-            else line.strip()
+            line.strip()
             for line in reqs_file.readlines()
             if not (
                 re.match(r"\s*#", line) or re.match("-e", line) or re.match("-r", line)
