+++ nucliadb_models/requirements.txt
@@ -1,2 +1,2 @@
 pydantic>=1.8.2,<2.0
-nucliadb-protos
+nucliadb-protos>=2.24.0
