+++ nucliadb_node/requirements.txt
@@ -7,3 +7,5 @@ grpcio-tools>=1.44.0
 grpcio-testing>=1.44.0
 grpcio-reflection>=1.44.0
 nucliadb-telemetry>=1.15.0
+nucliadb-protos>=2.24.0
+nucliadb-utils[storages,cache,fastapi]>=2.24.0
