+++ nucliadb_protos/python/setup.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 
 _dir = Path(__file__).resolve().parent
-VERSION = _dir.parent.parent.joinpath("VERSION").open().read().strip()
+VERSION = _dir.joinpath("VERSION").open().read().strip()
 README = _dir.joinpath("README.rst").open().read()
