+++ nucliadb_sdk/requirements.txt
@@ -1,6 +1,6 @@
 httpx
 requests
-nucliadb-models
+nucliadb-models>=2.24.0
 urllib3<1.27,>=1.21.1
 orjson
 pydantic>=1.10.9,<2.0
