+++ nucliadb_telemetry/VERSION
@@ -1,1 +1,1 @@
-1.16.0
+1.16.1
