+++ nucliadb_utils/setup.py
@@ -6,17 +6,14 @@ from typing import List
 from setuptools import find_packages, setup
 
 _dir = Path(__file__).resolve().parent
-VERSION = _dir.parent.joinpath("VERSION").open().read().strip()
+VERSION = _dir.joinpath("VERSION").open().read().strip()
 README = _dir.joinpath("README.md").open().read()
 
 
 def load_reqs(filename):
     with open(filename) as reqs_file:
         return [
-            # pin nucliadb-xxx to the same version as nucliadb
-            line.strip() + f"=={VERSION}"
-            if line.startswith("nucliadb-") and "=" not in line
-            else line.strip()
+            line.strip()
             for line in reqs_file.readlines()
             if not (
                 re.match(r"\s*#", line) or re.match("-e", line) or re.match("-r", line)
